requirements.txt
Django==1.8.1

#Admin
django-filebrowser==3.5.7
django-grappelli==2.6.4
feedparser==5.1.3
Pillow==2.8.1
django-localflavor==1.1

#Static hosting
django-pipeline==1.5.0
libsass==0.8.2
cssmin==0.2.0
slimit==0.8.1
jsmin==2.1.1
manifesto==0.5

#DB
psycopg2==2.6

#User Accounts
django-registration-redux==1.2
python-social-auth==0.2.9


#Utils
django-annoying==0.8.1
django-extensions==1.5.5
django-model-utils==2.2
django-braces==1.8.0
#TODO remove unipath
git+https://github.com/mikeorr/Unipath.git#egg=unipath
pytz==2015.4
easy-thumbnails==2.2
django-taggit==0.14.0


#REDIS CACHING
hiredis==0.2.0
redis==2.10.3
django-redis-cache==0.13.1
django-redis-sessions==0.4.0

## AWS
boto==2.38.0
django-storages-redux==1.2.3

#HEROKU
dj-database-url==0.3.0
#dj-static==0.0.5
-e git+https://github.com/gone/dj-static.git@5da00529c033afa8e65aef13555b9eaec6dfc388#egg=dj-static
django-toolbelt==0.0.1
ipython==1.1.0
waitress==0.8.9
djrill==1.4.0
gunicorn==19.3.0

#cool stuff
django-reversion==1.8.6

#OPTIONAL TOOLS
#whoosh
#django-haystack


#Money
#stripe==1.12.0
pip-accel==0.31
